+++ get_next_line_utils_bonus.c
@@ -6,15 +6,15 @@
 /*   By: yousenna <yousenna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/07 13:09:34 by yousenna          #+#    #+#             */
-/*   Updated: 2025/12/07 16:39:58 by yousenna         ###   ########.fr       */
+/*   Updated: 2025/12/11 22:50:56 by yousenna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "get_next_line.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	if (s == NULL)
